refactor: move writing out of config to consumer
