+++ lib/signs_ui/application.ex
@@ -13,6 +13,7 @@ defmodule SignsUi.Application do
       {Phoenix.PubSub, name: SignsUi.PubSub},
       SignsUiWeb.Endpoint,
       SignsUi.Config.State,
+      SignsUi.Config.Writer,
       {SignsUi.Signs.State, [name: SignsUi.Signs.State]},
       SignsUi.Config.Expiration,
       SignsUi.RefreshTokenStore,
